Revert "Rename: Move Header component"
This reverts commit 9592bb14d72fb63bf1ca618c505b1900f4eb842f.
