+++ src/app/category/page.tsx
@@ -2,7 +2,8 @@
 'use client';
 import React, { useEffect, useState } from 'react';
 
-import { Header, HeaderNav } from '@/components/common/header';
+import Header from '@/components/common/Header';
+import HeaderNav from '@/components/common/HeaderNav';
 
 // import InfiniteScroll from 'react-infinite-scroll-component';
 import Dropdown from '../../components/category/Dropdown';
@@ -261,7 +262,9 @@ const CategoryPage = () => {
   return (
     <div className='flex flex-col items-center justify-center bg-white pt-[3rem]'>
       <Header>
-        <HeaderNav showBack showCart showMyPage />
+        <HeaderNav showBack showCart showMyPage>
+          {' '}
+        </HeaderNav>
       </Header>
       <h1 className='m-5 text-lg'>어디로 갈까요?</h1>
       <Dropdown
