+++ src/app/detail/[id]/page.tsx
@@ -1,11 +1,12 @@
 import Image from 'next/image';
 import React from 'react';
 
-import { Header, HeaderNav } from '@/components/common/header';
 import {
   Carousel,
   CartButton,
   DatePicker,
+  Header,
+  HeaderNav,
   KakaoMap,
   PersonInput,
   ReservationButton,
