--- src/components/common/header/index.ts
@@ -1,4 +1,0 @@
-import Header from './Header';
-import HeaderNav from './HeaderNav';
-
-export { Header, HeaderNav };
